+++ spring-integration-jmx/src/main/java/org/springframework/integration/jmx/MBeanAttributeFilter.java
@@ -16,21 +16,15 @@
 
 package org.springframework.integration.jmx;
 
-import javax.management.ObjectName;
-
 /**
  * @author Stuart Williams
+ *
  * @since 3.0
  *
+ * @deprecated since 7.0 in favor {@link org.springframework.integration.jmx.inbound.MBeanAttributeFilter}
  */
+@Deprecated(forRemoval = true, since = "7.0")
 @FunctionalInterface
-public interface MBeanAttributeFilter {
-
-	/**
-	 * @param objectName The object name.
-	 * @param attributeName The attribute name.
-	 * @return true if the attribute passes the filter.
-	 */
-	boolean accept(ObjectName objectName, String attributeName);
+public interface MBeanAttributeFilter extends org.springframework.integration.jmx.inbound.MBeanAttributeFilter {
 
 }
