+++ spring-integration-jmx/src/main/java/org/springframework/integration/jmx/inbound/MBeanObjectConverter.java
@@ -14,17 +14,25 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.jmx;
+package org.springframework.integration.jmx.inbound;
+
+import javax.management.MBeanServerConnection;
+import javax.management.ObjectInstance;
 
 /**
  * @author Stuart Williams
  *
- * @since 3.0
+ * @since 7.0
  *
- * @deprecated since 7.0 in favor {@link org.springframework.integration.jmx.inbound.DefaultMBeanAttributeFilter}
  */
-@Deprecated(forRemoval = true, since = "7.0")
-public class DefaultMBeanAttributeFilter
-		extends org.springframework.integration.jmx.inbound.DefaultMBeanAttributeFilter {
+@FunctionalInterface
+public interface MBeanObjectConverter {
+
+	/**
+	 * @param connection The connection.
+	 * @param instance The instance.
+	 * @return The mapped object instance.
+	 */
+	Object convert(MBeanServerConnection connection, ObjectInstance instance);
 
 }
