+++ spring-integration-jmx/src/main/java/org/springframework/integration/jmx/inbound/NamedFieldsMBeanAttributeFilter.java
@@ -14,25 +14,41 @@
  * limitations under the License.
  */
 
-package org.springframework.integration.jmx;
+package org.springframework.integration.jmx.inbound;
+
+import java.util.Arrays;
+
+import javax.management.ObjectName;
+
+import org.springframework.util.Assert;
 
 /**
  * @author Stuart Williams
  * @author Artem Bilan
  *
- * @since 3.0
+ * @since 7.0
  *
- * @deprecated since 7.0 in favor {@link org.springframework.integration.jmx.inbound.NamedFieldsMBeanAttributeFilter}
  */
-@Deprecated(forRemoval = true, since = "7.0")
-public class NamedFieldsMBeanAttributeFilter
-		extends org.springframework.integration.jmx.inbound.NamedFieldsMBeanAttributeFilter {
+public class NamedFieldsMBeanAttributeFilter implements MBeanAttributeFilter {
+
+	private final String[] namedFields;
 
 	/**
 	 * @param namedFields The named fields that should pass the filter.
 	 */
 	public NamedFieldsMBeanAttributeFilter(String... namedFields) {
-		super(namedFields);
+		Assert.notNull(namedFields, "'namedFields' must not be null");
+		this.namedFields = Arrays.copyOf(namedFields, namedFields.length);
+	}
+
+	@Override
+	public boolean accept(ObjectName objectName, String attributeName) {
+		for (String namedField : this.namedFields) {
+			if (namedField.equals(attributeName)) {
+				return true;
+			}
+		}
+		return false;
 	}
 
 }
