+++ spring-integration-jmx/src/main/java/org/springframework/integration/jmx/outbound/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Outbound channel adapters for JMX support.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.integration.jmx.outbound;
