+++ spring-integration-jmx/src/test/java/org/springframework/integration/jmx/config/MBeanTreePollingChannelAdapterParserTests-context.xml
@@ -33,7 +33,7 @@
 			channel="channel2"
 			auto-startup="false">
 		<si:poller max-messages-per-poll="1" fixed-rate="5000"/>
-		<bean class="org.springframework.integration.jmx.DefaultMBeanObjectConverter"/>
+		<bean class="org.springframework.integration.jmx.inbound.DefaultMBeanObjectConverter"/>
 	</jmx:tree-polling-channel-adapter>
 
 	<si:channel id="channel2">
@@ -92,7 +92,7 @@
 		<si:poller max-messages-per-poll="1" fixed-rate="5000"/>
 	</jmx:tree-polling-channel-adapter>
 
-	<bean id="converter" class="org.springframework.integration.jmx.DefaultMBeanObjectConverter" />
+	<bean id="converter" class="org.springframework.integration.jmx.inbound.DefaultMBeanObjectConverter" />
 
 	<si:channel id="channel6">
 		<si:queue/>
